+++ native_library/android/GoogleMapsANE/app/src/main/java/com/tuarua/GoogleMapsANE.java
@@ -54,6 +54,7 @@ public class GoogleMapsANE implements FREExtension {
             ,"zoomIn"
             ,"zoomOut"
             ,"zoomTo"
+            ,"scrollBy"
             ,"setAnimationDuration"
             ,"requestPermissions"
             ,"capture"